Fix ConnectionManager subscription management
Since we just used `.do` with `.tap`, we lost the relationship between
the connection manager and the connections, and thus would not tear them
down when the connection pipeline was stopped.

It'd be nice to have a graceful stop too, but this change is also a
prerequisite for achieving that.

This fixes https://github.com/croservices/cro-websocket/issues/22 which
was a result of stopping the server not resulting in the termination of
connection pipelines that were ongoing; it was just much more noticeable
with Websockets, since the connections are long-lived.