Workaround to allow setting TCP_NODELAY
This is the first part (and also the most likely to need iteration) of
the fixes for https://github.com/croservices/cro/issues/129 .

Until https://github.com/MoarVM/MoarVM/issues/1229 is fixed and the
functionality is available directly in rakudo-moar, we need to use a
workaround to allow setting TCP_NODELAY using NativeCall instead.

This commit introduces Cro::TCP::NoDelay, which exports a sub that
enables setting this on any socket that supports .native-descriptor
(which both TCP and TLS async sockets do) on operating systems that
support the `setsockopt` call.

There are two portability concerns with this commit:

1. This commit assumes TCP_NODELAY == 1 and `int` is 32 bits.  I'm not
   sure how much of a problem this actually is.
2. There is no fallback for operating systems (or Raku VMs) that don't
   support calling `setsockopt` via NativeCall.

Help hereby requested to resolve issue 2, and determine whether issue 1
is a real problem for us.