Skip nodelay on Windows
Since the NativeCall code has not been tested on Windows (I don't have a
Windows system to test with), just silently skip the TCP_NODELAY code on
those distros.  Since the cost of this check could be paid on every socket
setup, cache the slow $*DISTRO.is-win lookup (it's not going to change in
the middle of runtime, afterall).