Test Connector and Listener with all nodelay combos
Run a simple 3-message send/reply sequence for every combination of client
Cro::TCP::Connector and server Cro::TCP::Listener nodelay options (without
the option, with it off, and with it on).

NOTE: This commit includes a (very short) sleep that doesn't seem like it
should be needed -- but if it isn't there, the test code will often deadlock
as soon as the first client message has been emitted.  My best guess as to
what causes this is the way that Cro::Connector.establish() waits for the
connection Promise before setting up the `whenever` that actually relays
messages, and thus that the Cro::TCP::Connector isn't actually ready for
a very short period.  If true however, this seems like a bug; I don't think
user code should be able to deadlock the connector in this way.