Fix IRI to URI conversion
If we have a valid IRI then it already has all percent encoding that a
URI would need for things in ASCII range, and atop of that needs
anything outside of ASCII range to be percent-encoded in order to form a
valid URI. Thus explicitly pick out the non-ASCII range when deciding
what to percent encode. Also, since this function is perhaps not very
useful in general, make it private to the Cro::Iri class (I checked it
doesn't seem to be used anywhere else).

Fixes https://github.com/croservices/cro-core/issues/35.
Fixes https://github.com/croservices/cro-http/issues/154.